+++ packages/shikiji-transformers/src/shared/highlight-word.ts
